functions refined for buttons

+++ livestock.cpp
@@ -20,7 +20,7 @@ int Livestock::Build(){
                     total_storage = 2;
                     user->Set_experience(user->Get_experience()+10);
 
-                    // timer
+                    building_timer = 5;
 
                     return 5;
                 }
@@ -34,22 +34,30 @@ int Livestock::Build(){
 }
 int Livestock::Upgrade() {
     // return values :
-    // 1 == limit for upgrade dou to user's level
-    // 2 == not enough coins
-    // 3 == not enough nails
-    // 4 == timer set for upgrade
+    // 1 == livestock is not built yet
+    // 2 == to upgrade you have to wait until feeding timer finishes and milks get collected
+    // 3 == limit for upgrade dou to user's level
+    // 4 == not enough coins
+    // 5 == not enough nails
+    // 6 == timer set for upgrade
 
-    if(user->Get_level()>=5){ // not limit for upgrade dou to user's level
-        if(user->Get_coin()>=15){ // enough coins
-            if(store->Get_object(2)>=2){ // enough nails
-                user->Set_coin(user->Get_coin()-15);
-                store->Delete(2,2);
-                level++;
-                total_storage *= 2;
-                user->Set_experience(user->Get_experience()+6);
+    if(building_status == 2){ // already built
+        if(feeding_status == 0){ // not fed and no product to collect
+            if(user->Get_level()>=5){ // not limit for upgrade dou to user's level
+                if(user->Get_coin()>=15){ // enough coins
+                    if(store->Get_object(2)>=2){ // enough nails
+                        user->Set_coin(user->Get_coin()-15);
+                        store->Delete(2,2);
+                        level++;
+                        total_storage *= 2;
+                        user->Set_experience(user->Get_experience()+6);
 
-                // timer;
+                        upgrade_timer = 5;
 
+                        return 6;
+                    }
+                    return 5;
+                }
                 return 4;
             }
             return 3;
@@ -65,22 +73,29 @@ int Livestock::Feed() {
     // 3 == product is ready. to feed cows first collect milk
     // 4 == not enough alfalfa
     // 5 == timer set for products
-    if(used_storage == 0){ // empty livestock
-        return 1;
+    // 6 == livestock is not built yet
+    if(building_status != 2){
+        return 6;
     }
-    else if(feeding_status == 1){ // already fed
-        return 2;
-    }
-    else if(feeding_status == 2){ // product is ready
-        return 3;
-    }
-    if(store->Delete(3,2*used_storage)){ // enough alfalfa
-        user->Set_experience(user->Get_experience()+(3*used_storage));
-        feeding_status = 1;
+    else{
+        if(used_storage == 0){ // empty livestock
+            return 1;
+        }
+        else if(feeding_status == 1){ // already fed
+            return 2;
+        }
+        else if(feeding_status == 2){ // product is ready
+            return 3;
+        }
+        if(store->Delete(3,2*used_storage)){ // enough alfalfa
+            user->Set_experience(user->Get_experience()+(3*used_storage));
+            feeding_status = 1;
+            feeding_timer = 3;
 
-        return 5;
+            return 5;
+        }
+        return 4;
     }
-    return 4;
 }
 
 bool Livestock::Add(int type, int amount) {
@@ -105,16 +120,22 @@ int Livestock::Collect() {
     // 1 == not product to collect
     // 2 == not enough space in store
     // 3 == collected successfully
-    if(feeding_status != 2){ // product is not ready yet
-        return 1;
+    // 4 == livstock is not built yet
+    if(building_status != 2){
+        return 4;
     }
     else{
-        if(store->Add(5,used_storage)){ // enough space in store
-            user->Set_experience(user->Get_experience()+5);
-            feeding_status = 0;
-
-            return 3;
+        if(feeding_status != 2){ // product is not ready yet
+            return 1;
         }
-        return 2;
+        else{
+            if(store->Add(5,used_storage)){ // enough space in store
+                user->Set_experience(user->Get_experience()+5);
+                feeding_status = 0;
+
+                return 3;
+            }
+            return 2;
+        }
     }
 }
